Refactor: Move GetFalconImage method from common to node specific module
It is in fact not at all that common.
