+++ pkg/common/funcs.go
@@ -1,15 +1,12 @@
 package common
 
 import (
-	"context"
 	"encoding/base64"
-	"fmt"
 	"regexp"
 	"strconv"
 	"strings"
 
 	falconv1alpha1 "github.com/crowdstrike/falcon-operator/apis/falcon/v1alpha1"
-	"github.com/crowdstrike/falcon-operator/pkg/registry/falcon_registry"
 	sprigcrypto "github.com/crowdstrike/falcon-operator/pkg/sprig"
 	"k8s.io/apimachinery/pkg/version"
 	"k8s.io/client-go/kubernetes"
@@ -23,33 +20,6 @@ func InitContainerArgs() []string {
 			" && " +
 			"touch " + FalconStoreFile,
 	}
-}
-
-func GetFalconImage(ctx context.Context, nodesensor *falconv1alpha1.FalconNodeSensor) (string, error) {
-	if nodesensor.Spec.Node.ImageOverride != "" {
-		return nodesensor.Spec.Node.ImageOverride, nil
-	}
-
-	if nodesensor.Spec.FalconAPI == nil {
-		return "", fmt.Errorf("Missing falcon_api configuration")
-	}
-
-	cloud, err := nodesensor.Spec.FalconAPI.FalconCloud(ctx)
-	if err != nil {
-		return "", err
-	}
-	imageUri := falcon_registry.ImageURINode(cloud)
-
-	registry, err := falcon_registry.NewFalconRegistry(ctx, nodesensor.Spec.FalconAPI.ApiConfig())
-	if err != nil {
-		return "", err
-	}
-	imageTag, err := registry.LastNodeTag(ctx, nil)
-	if err != nil {
-		return "", err
-	}
-
-	return fmt.Sprintf("%s:%s", imageUri, imageTag), nil
 }
 
 func FalconSensorConfig(falconsensor *falconv1alpha1.FalconSensor, cid string) map[string]string {
